Improve block creation (nicer interface, columns)
