+++ Sources/NotionSwift/Models/Blocks/ReadBlock.swift
@@ -1,13 +1,8 @@
 //
-//  Created by Wojciech Chojnacki on 22/05/2021.
+//  Created by Wojciech Chojnacki on 21/03/2022.
 //
 
 import Foundation
-
-/// Just a placeholder entity. User ReadBlock / WriteBlock structs.
-public enum Block {
-    public typealias Identifier = EntityIdentifier<Block, UUIDv4>
-}
 
 /// Block object from a server response
 public struct ReadBlock: CustomStringConvertible {
@@ -56,40 +51,7 @@ public struct ReadBlock: CustomStringConvertible {
     }
 }
 
-/// Block object used in adding a new content
-public struct WriteBlock {
-    public let type: BlockType
-    public let hasChildren: Bool
-
-    public init(type: BlockType, hasChildren: Bool = false) {
-        self.type = type
-        self.hasChildren = hasChildren
-    }
-}
-
-public struct UpdateBlock {
-    public let value: BlockType?
-    public let archived: Bool?
-
-    public init(value: BlockType?, archived: Bool? = nil) {
-        self.value = value
-        self.archived = archived
-    }
-}
-
 // MARK: - Codable
-
-extension Block {
-    enum CodingKeys: String, CodingKey {
-        case id
-        case type
-        case archived
-        case createdTime = "created_time"
-        case lastEditedTime = "last_edited_time"
-        case hasChildren = "has_children"
-        case object
-    }
-}
 
 extension ReadBlock: Decodable {
     public init(from decoder: Decoder) throws {
@@ -101,25 +63,6 @@ extension ReadBlock: Decodable {
         type = try BlockType(from: decoder)
         hasChildren = try container.decode(Bool.self, forKey: .hasChildren)
         children = []
-    }
-}
-
-extension WriteBlock: Encodable {
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: Block.CodingKeys.self)
-        try type.encode(to: encoder)
-        try container.encode(hasChildren, forKey: .hasChildren)
-        try container.encode("block", forKey: .object)
-    }
-}
-
-extension UpdateBlock: Encodable {
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: Block.CodingKeys.self)
-        if let value = value {
-            try value.encode(to: encoder)
-        }
-        try container.encodeIfPresent(archived, forKey: .archived)
     }
 }
